==> CODE/TRANSFORM/01_entity_details_mover.ipynb <==
import pandas as pd
import numpy as np

RAW_FOLDER = '../../DATA/RAW/'
TRANSFORM_FOLDER = '../../DATA/TRANSFORM/'

## VAT CANCELLATION PUBLICATION DATE

dat = pd.read_csv(RAW_FOLDER + 'zrusenie_registracie_dph/final/zrusenie_registracie_dph.csv')
dat.columns = ['ic_dph', 'entity_name', 'entity_city', 'entity_zip', 'entity_address', 'vat_cancellation_year', 'vat_cancellation_publication_date', 'entity_ico']
dat.query('entity_ico == entity_ico').to_csv(TRANSFORM_FOLDER + 'entity_details/vat_registration_cancellation.csv', index = False)

## INSOLVENCY REGISTER

dat = pd.read_csv(RAW_FOLDER + 'register_upadcov/final/register_upadcov.csv')
dat.columns = ['proceeding_start', 'proceeding_label', 'court', 'entity_name',
       'proceeding_start_date', 'entity_ico', 'proceeding_type', 'proceeding_month']

dat.query('entity_ico == entity_ico').to_csv(TRANSFORM_FOLDER + 'entity_details/insolvency_register.csv', index = False)

## FINANCIAL STATEMENTS REGISTER

dat = (pd.concat(
    [
        pd.read_csv('../../DATA/RAW/register_uz_apr23/preprocessed/entity_details_121.csv'),
        pd.read_csv('../../DATA/RAW/register_uz_apr23/preprocessed/entity_details_112.csv')
        ], axis = 0))

dat.columns = ['id', 'annual_reports_list', 'financial_statements_list', 'entity_ico', 'entity_dic',
       'entity_name', 'entity_city', 'entity_address', 'entity_zip', 'modified_date',
       'establishment_date', 'legal_form_code', 'org_size_code', 'entity_county', 'entity_registerred_office',
       'data_source', 'ownership_category_code', 'entity_region', 'sknace_code', 'consolidated',
       'cancellation_date', 'entity_state', 'entity_name_RO', 'entity_city_RO', 'legal_form_RO',
       'establishment_date_RO', 'cancellation_date_RO', 'entity_id']


dat =  dat[['id', 'annual_reports_list', 'financial_statements_list', 'entity_ico',
       'entity_name', 'establishment_date', 'legal_form_code', 'org_size_code', 'entity_county',
       'data_source', 'ownership_category_code', 'entity_region', 'sknace_code', 'consolidated',
       'cancellation_date', 'entity_state', 'legal_form_RO',
       'establishment_date_RO', 'cancellation_date_RO', 'entity_id']]


dat.to_csv(TRANSFORM_FOLDER + 'entity_details/financial_statements_register_entity_details.csv', index = False)
