==> CODE/TRANSFORM/02_financial_statements_mover.ipynb <==
import pandas as pd
import numpy as np

import ast

RAW_FOLDER = '../../DATA/RAW/'
TRANSFORM_FOLDER = '../../DATA/TRANSFORM/financial_statements/'

statements_table_112 = pd.read_csv('../../DATA/RAW/register_uz_apr23/preprocessed/financial_statements_112.csv')
statements_table_121 = pd.read_csv('../../DATA/RAW/register_uz_apr23/preprocessed/financial_statements_121.csv')


statements_table_112.rename(columns={'id': 'financial_statement_id', 'idUJ': 'entity_id'}, inplace=True)
statements_table_121.drop(['entity_id'], axis = 1, inplace=True)
statements_table_121.rename(columns={'idUJ': 'entity_id'}, inplace=True)

statements_table = (pd.concat(
    [
        statements_table_112,
        statements_table_121
        ], axis = 0))

statements_table_112 = None
statements_table_121 = None

(statements_table
 .rename(
     columns=
     {
        #  'id': 'financial_statement_id',
         'obdobieOd': 'starting_from',
         'obdobieDo': 'until',
         'datumZostaveniaK': 'assembly_date',
        #  'idUJ': 'entity_id',
         'idUctovnychVykazov': 'reports_list',
         'typ': 'type',
         'zdrojDat': 'data_source',
         'ico': 'entity_ico',
         'dic': 'entity_dic',
         'nazovUJ': 'entity_name',
         'datumPoslednejUpravy': 'modified',
         'datumPodania': 'submission_date',
         'datumZostavenia': 'prepared_date',
         'datumSchvalenia': 'approved_date',
         'datumPrilozeniaSpravyAuditora': 'auditors_report_attachement_date',
         'konsolidovana': 'consolidated',
         'nazovFondu': 'fund_name',
         'leiKod': 'lei_code'
     }, inplace=True))

statements_table


def period_length(x):
    return (x.period_end_year - x.year) * 12 + x.period_end_month - x.month + 1

statements_table['year'] = statements_table['starting_from'].apply(lambda x: int(x[:4]))
statements_table['month'] = statements_table['starting_from'].apply(lambda x: int(x[5:]))
statements_table['period_end_year'] = statements_table['until'].apply(lambda x: int(x[:4]))
statements_table['period_end_month'] = statements_table['until'].apply(lambda x: int(x[5:]))

statements_table['period_length'] = statements_table.apply(period_length, axis = 1)
statements_table

statements_table.query('year >= 2013', inplace=True)

statements_table.to_csv(TRANSFORM_FOLDER+'financial_statements_table.csv', index = False)

statements_reports_table = statements_table[['financial_statement_id', 'reports_list']]

statements_reports_records = []

k = 0
for index, row in statements_reports_table.query('reports_list == reports_list').iterrows():
    reports_list = ast.literal_eval( row['reports_list'] )
    if not isinstance(reports_list, list):
        print(row['reports_list'], type(reports_list))
        continue
    for i, report in enumerate(reports_list):
        statements_reports_records.append([row['financial_statement_id'], report, i + 1])
    
statements_reports = pd.DataFrame(statements_reports_records, columns=['financial_statement_id', 'financial_report_id', 'financial_report_order_number'])

statements_reports
    


statements_reports.to_csv(TRANSFORM_FOLDER+'financial_statements_reports_table.csv', index = False)
