==> CODE/TRANSFORM/03_financial_reports_mover.ipynb <==
import pandas as pd
import numpy as np

import ast

files = [
    'financial_reports_112_01',
    'financial_reports_112_02',
    'financial_reports_112_03',
    'financial_reports_112_04',
    'financial_reports_112_05',
    'financial_reports_112_06',
    'financial_reports_121'
]

columns_names = {
    'prilohy': 'attachments',
    'id': 'financial_report_id',
    'idUctovnejZavierky': 'financial_statement_id',
    'obsah': 'content',
    'pristupnostDat': 'data_access',
    'idSablony': 'template_id',
    'datumPoslednejUpravy': 'modified',
    'zdrojDat': 'source',
    'kodDanovehoUradu': 'tax_office_code'}


df = pd.DataFrame()

for file in files:
    print(file)
    temp = pd.read_csv(f'../../DATA/RAW/register_uz_apr23/preprocessed/{file}.csv')
    if file == 'financial_reports_121': temp.drop(columns='idUctovnejZavierky', inplace=True)
    temp.rename(columns=columns_names, inplace=True)

    df = pd.concat([df, temp], axis = 0)
    temp = None

# df = df.merge(templates, on = 'template_id', how = 'left')
df

df.query('template_id in [687, 699] and content == content and content != "{}"', inplace=True)

display(df['template_id'].value_counts())

df.to_csv('../../DATA/TRANSFORM/financial_statements/financial_reports_table.csv', index = False)
