==> CODE/TRANSFORM/04_financial_reports_content_parsing.ipynb <==
from HELPERS.ReportContentParser import ReportContentParser

import pandas as pd
import numpy as np

import ast


pd.set_option('display.max_rows', 100)

reports_table = pd.read_csv('../../DATA/TRANSFORM/financial_statements/financial_reports_table.csv')
statements_table = pd.read_csv('../../DATA/TRANSFORM/financial_statements/financial_statements_table.csv')

statements_table = statements_table[['financial_statement_id', 'entity_id', 'entity_ico', 'year', 'month']]

reports_table = reports_table.merge(statements_table, on = 'financial_statement_id', how = 'left')
statements_table = None

parser = ReportContentParser()

records = []

size = reports_table.shape[0]

i = 0
for report in reports_table.itertuples():
    i += 1
    if i % 1000 == 0:
        print(f'{i}/{size}')
    # if i > 10:
    #     break

    financial_report_id = report.financial_report_id

    values = parser.parse(report.template_id, report.content)
    if len(values) == 0:
        continue

    records.append([financial_report_id, report.template_id] + values)

columns_names = [
    'financial_report_id',
    'financial_report_template_id',
    'total_assets',
    'non_current_assets',
    'intangible_assets',
    'total_non_current_tangible_assets',
    'land_and_buildings',
    'separate_movable_assets_and_sets_of_movable_assets',
    'other_long_term_tangible_assets',
    'adjustment_item_for_acquired_assets',
    'total_long_term_financial_assets',
    'bank_balances_with_maturity_longer_than_one_year',
    'current_assets',
    'inventories',
    'non_current_claims',
    'current_claims',
    'trade_receivables',
    'social_security_tax_receivables_and_grants',
    'financial_assets',
    'other_financial_accounts',
    'cash_and_bank_balances',
    'total_owners_equity_and_liabilities',
    'owners_equity',
    'share_capital',
    'share_capital_and_changes_in_share_capital',
    'claims_for_subscribed_owners_equity',
    'capital_funds',
    'valuation_differences',
    'undivided_profit_or_unpaid_loss_from_previous_years',
    'profit_for_the_reporting_period_after_tax',
    'liabilities',
    'long_term_liabilities_excluding_reserves_and_loans',
    'long_term_reserves',
    'long_term_bank_loans',
    'total_short_term_liabilities_excluding_reserves_loans_and_advances',
    'short_term_trade_liabilities',
    'liabilities_to_employees_and_social_insurance',
    'tax_liabilities_and_grants',
    'short_term_reserves',
    'current_bank_loans',
    'short_term_financial_assistance',
    'total_operating_revenue',
    'sales_revenue',
    'revenue_from_the_sale_of_own_products_and_services',
    'change_in_intra_organizational_inventories',
    'activation',
    'revenue_from_the_sale_of_long_term_intangible_assets_long_term_tangible_assets_and_materials',
    'other_revenues_from_operating_activities',
    'total_operating_costs',
    'costs_incurred_to_acquire_sold_goods',
    'consumption_of_materials_energy_and_other_non_inventory_supplies',
    'services',
    'personnel_costs',
    'taxes_and_fees',
    'depreciation_and_adjustments_to_long_term_intangible_assets_and_long_term_tangible_assets',
    'remaining_cost_of_sold_long_term_assets_and_materials',
    'adjustments_to_receivables',
    'other_costs_of_operating_activities',
    'operating_result_from_operating_activities',
    'value_added',
    'total_financial_revenues',
    'revenue_from_the_sale_of_securities_and_shares',
    'revenue_from_long_term_financial_assets',
    'revenue_from_short_term_financial_assets',
    'interest_income',
    'exchange_gains',
    'other_revenues_from_financial_activities',
    'total_financial_costs', 'sale_of_securities_and_shares',
    'costs_of_short_term_financial_assets',
    'adjustments_to_financial_assets',
    'interest_expenses',
    'exchange_losses',
    'other_costs_of_financial_activities',
    'financial_result',
    'result_of_operations_for_the_reporting_period_before_tax',
    'income_tax',
    'transfer_of_shares_in_the_result_to_shareholders',
    'result_of_operations_for_the_reporting_period_after_tax'
]

df = pd.DataFrame(records, columns=columns_names)

display(df)



df.to_csv('../../DATA/MODEL/raw_primary_values_df.csv', index = False)

# df = pd.read_csv('../../DATA/MODEL/raw_primary_values_df.csv')

financial_statements_reports = reports_table[['financial_report_id', 'financial_statement_id', 'template_id', 'entity_ico', 'year']]
financial_statements_reports

reports_table = None

renaming = {
    # STRANA AKTIV
    'total_assets': 'total_assets',

    # A
    'non_current_assets': 'non_current_assets',
    'intangible_assets': 'non_current_intangible_assets',
    'total_non_current_tangible_assets': 'non_current_tangible_assets',
    'total_long_term_financial_assets': 'non_current_financial_assets',

    # B
    'current_assets': 'current_assets',
    'inventories': 'inventories',
    'non_current_claims': 'non_current_receivables',
    'current_claims': 'current_receivables',
    'financial_assets': 'financial_assets', # current financial assets + financial accounts

    'other_financial_accounts': 'current_financial_assets',
    'cash_and_bank_balances': 'cash_and_bank_balances',

    # STRANA PASIV
    
    # A
    'total_owners_equity_and_liabilities': 'total_equity_and_liabilities', # A
    'owners_equity': 'equity', # AI
    'share_capital': 'share_capital', # AII
    'capital_funds': 'share_premium', # AIII
    # missing other capital funds # AIV
    # missing other funds created from profit # AV
    # missing differences from evaluation # AVI
    'undivided_profit_or_unpaid_loss_from_previous_years': 'net_profit_of_previous_years', # AVII
    'profit_for_the_reporting_period_after_tax': 'net_profit_after_tax', # AVIII

    # B
    'liabilities': 'liabilities', # B
    'long_term_liabilities_excluding_reserves_and_loans': 'non_current_liabilities', # BI
    'long_term_reserves': 'long_term_provisions', # BII
    'long_term_bank_loans': 'long_term_bank_loans', # BIII
    'total_short_term_liabilities_excluding_reserves_loans_and_advances': 'current_liabilities', # BIV
    'short_term_reserves': 'short_term_provisions', # BV
    'current_bank_loans': 'current_bank_loans', # BVI
    'short_term_financial_assistance': 'short_term_financial_assistance', # BVII
    
    # VYKAZ ZISKOV A STRAT
    
    # OPERATING REVENUES
    'total_operating_revenue': 'operating_revenues',

    'sales_revenue': 'sales_from_the_merchandise',
    'revenue_from_the_sale_of_own_products_and_services': 'sales_from_the_own_products_and_services',    
    'revenue_from_the_sale_of_long_term_intangible_assets_long_term_tangible_assets_and_materials': 'sales_from_the_other',
    'other_revenues_from_operating_activities': 'other_revenues_from_operating_activities',

    # OPERATING COSTS
    'total_operating_costs': 'operating_costs',

    'costs_incurred_to_acquire_sold_goods': 'costs_of_merchandise_sold',


    'consumption_of_materials_energy_and_other_non_inventory_supplies': 'consumed_materials_energy_non_inventory_supplies',

    'services': 'services',
    'personnel_costs': 'personnel_costs',
    'taxes_and_fees': 'taxes_and_fees',
    'depreciation_and_adjustments_to_long_term_intangible_assets_and_long_term_tangible_assets': 'depreciation',

    'remaining_cost_of_sold_long_term_assets_and_materials': 'remaining_cost_of_sold_long_term_assets_and_materials',
    'adjustments_to_receivables': 'adjustments_to_receivables',
    'other_costs_of_operating_activities': 'other_costs_of_operating_activities',


    # OPERATING PROFIT/LOSS
    'operating_result_from_operating_activities': 'operating_result',
    
    # ADDED VALUE
    'value_added': 'value_added',


    'total_financial_revenues': 'revenues_from_financial_activities',
    'total_financial_costs': 'cost_of_financial_activities',
    
    # FINANCIAL ACTIVITIES PROFIT/LOSS
    'financial_result': 'result_from_fincancial_activities',

    # PROFIT LOSS FOR THE ACCOUNTING PERIOD BEFORE TAX
    'result_of_operations_for_the_reporting_period_before_tax': 'result_before_tax',
    'income_tax': 'income_tax',
    
    'result_of_operations_for_the_reporting_period_after_tax': 'result_after_tax'
}



df.rename(columns = renaming, inplace = True)

# Balance sheet primary values
df['TA'] = df['total_assets']
df['CA'] = df['current_assets']
df['CASH'] = df['financial_assets']
df['CC'] = df['current_receivables']
df['INV'] = df['inventories']

df['EQ'] = df['equity']
df['TL'] = df['liabilities']
df['NCL'] = df['non_current_liabilities']
df['CL'] = df['current_liabilities']

# Income statement primary values
df['CF_NETTO'] = df['depreciation'] + df['result_before_tax']
df['CF_SELFFIN'] = df['result_after_tax'] + df['depreciation'] + df['change_in_intra_organizational_inventories']

df['EAT'] = df['result_after_tax']
df['EBIT'] = df['result_before_tax']
df['SAL'] = df['operating_revenues']
df['OE'] = df['operating_result']
df['REV'] = df['operating_revenues'] + df['revenues_from_financial_activities']
df['VA'] = df['value_added']
df['COST'] = df['operating_costs']
df['INT'] = df['interest_income'] + df['interest_expenses']

df['LABOR'] = df['personnel_costs']


priamry_processed_values = ['CA','CASH','CL','EQ','NCL','TA','TL','CC','INV','CF_NETTO','CF_SELFFIN', 'REV', 'VA', 'OE' ,'EAT','EBIT','SAL','COST','INT', 'LABOR']


display(df)




df = df[['financial_report_id', 'financial_report_template_id'] + list(renaming.values()) + priamry_processed_values]

df.to_csv('../../DATA/MODEL/primary_features_df.csv', index = False)
