==> CODE/TRANSFORM/05_financial_ratios_transform.ipynb <==
import pandas as pd
import numpy as np


pd.set_option('display.max_columns', 130)
pd.set_option('display.max_rows', 100)

## Loading main dataset of primary metrics

df = pd.read_csv('../../DATA/MODEL/primary_features_df.csv')

priamry_processed_values = ['CA','CASH','CL','EQ','NCL','TA','TL','CC','INV','CF_NETTO','CF_SELFFIN', 'REV', 'VA', 'OE' ,'EAT','EBIT','SAL','COST','INT', 'LABOR']


df.query('SAL >= 0 and CA >= 0 and TL >= 0 and CASH >= 0 and NCL >= 0 and total_equity_and_liabilities == total_assets', inplace=True)

## Computing secondary metrics - financial ratios

def get_ratios(row):
    # Ukazovatele likvidity
    row['L1'] = ( row['CASH'] ) / ( row['CL'] ) if row['CL'] != 0 else None
    row['L2'] = ( row['CASH'] + row['CC'] ) / ( row['CL'] ) if row['CL'] != 0 else None
    row['L3'] = ( row['CASH'] + row['CC'] + row['INV'] ) / ( row['CL'] ) if row['CL'] != 0 else None
    row['NCL_CF'] = ( row['NCL'] ) / ( row['CF_NETTO'] ) if row['CF_NETTO'] != 0 else None
    row['CASH_TA'] = ( row['CASH'] ) / ( row['TA'] ) if row['TA'] != 0 else None

    # Ukazovatele aktivity
    row['INV_COGS'] = ( row['INV'] ) / ( row['costs_of_merchandise_sold'] ) if row['costs_of_merchandise_sold'] != 0 else None
    row['CC_SAL'] = ( row['CC'] ) / ( row['SAL'] ) if row['SAL'] != 0 else None
    row['TL_COST'] = row['TL'] / ( row['COST'] ) if row['COST'] > 0 else None
    row['CL_COST'] = row['CL'] / ( row['COST'] ) if row['COST'] > 0 else None
    row['SAL_TA'] = ( row['SAL'] ) / ( row['TA'] ) if row['TA'] != 0 else None

    # Ukazovatele zadlženosti
    row['EQ_TA'] = row['EQ'] / ( row['TA'] ) if row['TA'] > 0 else None
    row['EQ_TL'] = ( row['EQ'] ) / ( row['EQ'] + row['TL'] ) if ( row['EQ'] + row['TL'] ) != 0 else None
    row['CL_CC'] = ( row['CL'] ) / ( row['CC'] ) if row['CC'] != 0 else None    
    row['LTC_NCA'] = ( row['EQ'] + row['TL'] - row['CL'] ) / ( row['TA'] - row['CA'] )  if ( row['TA'] - row['CA'] ) != 0 else None
    row['TZ'] = ( row['TL'] - row['CASH'] - row['CC'] - row['INV'] ) / row['CF_NETTO']  if row['CF_NETTO'] != 0 else None
    row['CF_TL'] = ( row['CF_NETTO'] ) / ( row['TL'] ) if row['TL'] != 0 else None
    row['CL_TA'] = ( row['CL'] ) / ( row['TA'] ) if row['TA'] != 0 else None
    row['NCL_TA'] = ( row['NCL'] ) / ( row['TA'] ) if row['TA'] != 0 else None

    # Ukazovatele rentability
    row['EBT_REV'] = ( row['EBIT'] ) / ( row['REV'] ) if row['REV'] != 0 else None
    row['ROS'] = ( row['EAT'] ) / ( row['SAL'] ) if row['SAL'] != 0 else None
    row['EAT_REV'] = ( row['EAT'] ) / ( row['REV'] ) if row['REV'] != 0 else None
    row['ROA'] = ( row['EAT'] ) / ( row['TA'] ) if row['TA'] != 0 else None
    row['ROE'] = ( row['EAT'] ) / ( row['EQ'] ) if row['EQ'] != 0 else None
    row['CF_TA'] = ( row['CF_NETTO'] ) / ( row['TA'] ) if row['TA'] != 0 else None
    row['CF_SAL'] = ( row['CF_NETTO'] ) / ( row['SAL'] ) if row['SAL'] != 0 else None

    # Ukazovatele efektivity
    row['LABOR_EAT'] = row['personnel_costs'] / ( row['EAT'] ) if row['EAT'] != 0 else None    
    row['LABOR_REV'] = row['personnel_costs'] / ( row['REV'] ) if row['REV'] != 0 else None
   

    return row


records = []
k = 0
size = df.shape[0]

for row_dict in df.to_dict(orient="records"):
    # if k >= 100:
    #     break

    if (k + 1) % 50000 == 0:
        print(f'{k + 1} / {size}')

    k += 1

    records.append(get_ratios(row_dict))


df_extended = pd.DataFrame(records)

ratios = ['L1',
       'L2', 'L3', 'NCL_CF', 'CASH_TA', 'INV_COGS', 'CC_SAL', 'TL_COST',
       'CL_COST', 'SAL_TA', 'EQ_TA', 'EQ_TL', 'CL_CC', 'LTC_NCA', 'TZ',
       'CF_TL', 'CL_TA', 'NCL_TA', 'EBT_REV', 'ROS', 'EAT_REV', 'ROA', 'ROE',
       'CF_TA', 'CF_SAL', 'LABOR_EAT', 'LABOR_REV']



df = None

df_extended.columns

## Joining meta information (dimensions) about financial statements

statements_reports_table = pd.read_csv('../../DATA/TRANSFORM/financial_statements/financial_statements_reports_table.csv')
statements_table = pd.read_csv('../../DATA/TRANSFORM/financial_statements/financial_statements_table.csv')
entity_details_table = pd.read_csv('../../DATA/TRANSFORM/entity_details/financial_statements_register_entity_details.csv')

entity_details_table = entity_details_table[['id', 'establishment_date', 'cancellation_date', 'sknace_code', 'entity_state', 'legal_form_code', 'org_size_code',  'entity_name', 'ownership_category_code', 'consolidated']].rename(columns={'id': 'entity_id', 'consolidated':'entity_consolidated'})
statements_table = statements_table[['financial_statement_id', 'entity_id', 'entity_ico', 'year', 'month', 'period_length', 'consolidated', 'approved_date', 'type']]

df_extended.drop_duplicates(inplace=True)

display(df_extended.shape)

statements_reports_table

df_statements_extended = (df_extended
                          .merge(
                              statements_reports_table.drop_duplicates('financial_report_id'),
                              on = 'financial_report_id',
                              how = 'left'))

display(df_statements_extended.shape)

df_statements_extended = (df_statements_extended
                          .merge(
                              statements_table.drop_duplicates('financial_statement_id'),
                              on = 'financial_statement_id',
                              how = 'left'))

display(df_statements_extended.shape)

df_statements_extended = (df_statements_extended
                          .merge(
                              entity_details_table.drop_duplicates('entity_id'),
                              on = 'entity_id',
                              how = 'left'))

display(df_statements_extended.shape)

df_extended = None
statements_reports_table = None
statements_table = None
entity_details_table = None

df_statements_extended['period_length_check'] = df_statements_extended['period_length'].apply(lambda x: 1 if x == 12 else 0)
df_statements_extended['period_length_check'].value_counts()

# filtrujeme 2014 - 2021
# ak existuje viac zaznamov z rovnakeho roka, mesiaca pre jednu entitu, berieme tie, ktore maju spravnu dlzku, su schvalene a pri rovnosti vyssie TA+SAL

display(df_statements_extended.shape)

df_statements_extended = (df_statements_extended
                          .query('2014 <= year <= 2021')
                          .sort_values(['period_length_check', 'approved_date', 'TA', 'SAL'], ascending=[False, False, False, False])
                          .groupby(['entity_id', 'year', 'month']).first()
                          .reset_index())

df_statements_extended.drop(columns=['period_length_check'], inplace=True)

display(df_statements_extended.shape)

df_statements_extended['period_length'].value_counts().head(20)

df_statements_extended.query('period_length > 12')

## Joining SKNACE codebook

codebook_sknace_raw = pd.read_csv('../../DATA/RAW/register_uz_apr23/preprocessed/codebook_sk-nace.csv')
codebook_sknace_divisions_raw = pd.read_csv('../../DATA/RAW/register_uz_apr23/preprocessed/codebook_sk-nace_divisions_su.csv', sep= '|')


codebook_sknace = codebook_sknace_raw[['kod', 'nazov.sk']]
codebook_sknace.columns = ['sknace_code', 'sknace_group']

codebook_sknace['sknace_code'] = pd.to_numeric(codebook_sknace['sknace_code'])
codebook_sknace['sknace_division_code'] = codebook_sknace['sknace_code'].apply(lambda x: x//1000)

# display(codebook_sknace)
# display(codebook_sknace_divisions_raw)

codebook_sknace_divisions = codebook_sknace_divisions_raw[['code', 'officialTitle', 'note']]
codebook_sknace_divisions.columns = ['sknace_division_code', 'sknace_division_name', 'sknace_division']
codebook_sknace_divisions['sknace_division_code'] = pd.to_numeric(codebook_sknace_divisions['sknace_division_code'])

codebook_sknace = codebook_sknace.merge(codebook_sknace_divisions, on = 'sknace_division_code', how = 'left')

codebook_sknace



df_statements_extended.shape[0]

merged_df = df_statements_extended.merge(codebook_sknace[['sknace_code', 'sknace_division_name', 'sknace_division']], on = 'sknace_code', how = 'left')
merged_df['sknace_subcategory'] = merged_df['sknace_code'] // 1000

# display(merged_df['sknace_subcategory'].value_counts().to_frame())

merged_df.shape[0]

df_statements_extended = None

merged_df.to_csv('../../DATA/MODEL/financial_ratios_df.csv', index = False)
