==> CODE/TRANSFORM/06_issues_indicator_table.ipynb <==
import pandas as pd
import numpy as np

import os


pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 100)

TRANSFORM_FOLDER = '../../DATA/TRANSFORM/entity_details/'
FINAL_FOLDER = '../../DATA/FINAL'

financial_ratios = pd.read_csv('../../DATA/MODEL/financial_ratios_df.csv')

financial_ratios

## Joining financial issues indicators

files = [
 'insolvency_register.csv',
 'vat_registration_cancellation.csv']

data = dict()
data['insolvency_register'] = pd.read_csv(TRANSFORM_FOLDER + 'insolvency_register.csv')
data['vat_registration_cancellation'] = pd.read_csv(TRANSFORM_FOLDER + 'vat_registration_cancellation.csv')

data['insolvency_register']['proceeding_type'].value_counts()

insolvency_info = data['insolvency_register'].copy()
insolvency_info['proceeding_type'] = insolvency_info['proceeding_type'].apply(lambda x: 'ostatne_konania' if x not in ['konkurz', 'restrukturalizacia'] else x)
insolvency_info

insolvency_info.drop_duplicates(['entity_ico', 'proceeding_type', 'proceeding_month'], inplace=True)

insolvency_info = insolvency_info.sort_values('proceeding_month', ascending=False).groupby(['entity_ico', 'proceeding_type', 'proceeding_month']).tail(1)

insolvency_info.rename(columns={'proceeding_type': 'event_type', 'proceeding_month': 'year'}, inplace=True)

insolvency_info = insolvency_info[['entity_ico', 'year', 'event_type']]

insolvency_info['event_type'] = insolvency_info['event_type'].apply(lambda x: 'RU_' + x)
insolvency_info



insolvency_info = data['insolvency_register'].copy().drop_duplicates(['entity_ico', 'proceeding_month'])[['entity_ico', 'proceeding_month']]
insolvency_info.columns = ['entity_ico', 'year']
insolvency_info['event_type'] = 'RU_event'

insolvency_info


vat_registration_cancellation_info = data['vat_registration_cancellation'].copy()
# vat_registration_cancellation_info

vat_registration_cancellation_info['entity_ico'] = pd.to_numeric(vat_registration_cancellation_info['entity_ico'], errors = 'coerce')
vat_registration_cancellation_info.query('entity_ico == entity_ico', inplace=True)
vat_registration_cancellation_info = vat_registration_cancellation_info[['entity_ico', 'vat_cancellation_year']]
vat_registration_cancellation_info['vat_cancellation_year'] = pd.to_numeric(vat_registration_cancellation_info['vat_cancellation_year'])
vat_registration_cancellation_info['event_type'] = 'DPHZ_vat_registration_cancelled'
vat_registration_cancellation_info.rename(columns = {'vat_cancellation_year': 'year'}, inplace=True)
vat_registration_cancellation_info

valid_ico_list = list(financial_ratios['entity_ico'].unique())

entity_issues_events_long = (pd.concat(
    [
         insolvency_info,
         vat_registration_cancellation_info
    ], axis = 0))
entity_issues_events_long.query('entity_ico in @valid_ico_list and year >= 2014', inplace=True)

entity_issues_events_long['event_type'].value_counts()

entity_issues_events_long['indicator'] = 1
entity_issues_events_long.to_csv('../../DATA/TRANSFORM/issues_events.csv', index = False)


entity_issues_indicators_1y_off = entity_issues_events_long.copy()
entity_issues_indicators_2y_off = entity_issues_events_long.copy()
entity_issues_indicators_3y_off = entity_issues_events_long.copy()

entity_issues_indicators_1y_off['year'] = entity_issues_indicators_1y_off['year'] - 1
entity_issues_indicators_1y_off['event_type'] = entity_issues_indicators_1y_off['event_type'].apply(lambda x: x+'_1y_off')

entity_issues_indicators_2y_off['year'] = entity_issues_indicators_2y_off['year'] - 2
entity_issues_indicators_2y_off['event_type'] = entity_issues_indicators_2y_off['event_type'].apply(lambda x: x+'_2y_off')

entity_issues_indicators_3y_off['year'] = entity_issues_indicators_3y_off['year'] - 3
entity_issues_indicators_3y_off['event_type'] = entity_issues_indicators_3y_off['event_type'].apply(lambda x: x+'_3y_off')


entity_issues_indicators = (pd.concat(
    [
        entity_issues_events_long.copy(),
        entity_issues_indicators_1y_off,
        entity_issues_indicators_2y_off,
        entity_issues_indicators_3y_off
    ], axis=0
))

entity_issues_events = pd.pivot_table(entity_issues_indicators, index = ['entity_ico', 'year'], columns = 'event_type', values='indicator', fill_value=0)
entity_issues_events.reset_index(inplace=True)
entity_issues_events

ISSUES_TYPES = ['DPHZ_vat_registration_cancelled', 'RU_event']

def process_entity_issues_2y_off(row, issue_type):
        return max(row[issue_type+'_1y_off'], row[issue_type+'_2y_off'])

def process_entity_issues_3y_off(row, issue_type):
        return max(row[issue_type+'_1y_off'], row[issue_type+'_2y_off'], row[issue_type+'_3y_off'])

for issue_type in ISSUES_TYPES:
    entity_issues_events[issue_type+'_2y_off'] = entity_issues_events.apply(lambda x: process_entity_issues_2y_off(x, issue_type), axis = 1)
    entity_issues_events[issue_type+'_3y_off'] = entity_issues_events.apply(lambda x: process_entity_issues_3y_off(x, issue_type), axis = 1)

entity_issues_events

display(financial_ratios.shape[0])

financial_ratios = financial_ratios.merge(entity_issues_events, on = ['entity_ico', 'year'], how = 'left')
display(financial_ratios.shape[0])


financial_ratios_copy = None

financial_ratios['cancellation_year'] = pd.to_datetime(financial_ratios['cancellation_date'], errors='coerce').dt.year

def ruz_cancelled(row):
    if row.year >= row.cancellation_year: return 1
    return 0

def ruz_cancelled_1y_off(row):
    if row.year + 1 >= row.cancellation_year: return 1
    return 0

def ruz_cancelled_2y_off(row):
    if row.year + 2 >= row.cancellation_year: return 1
    return 0


def ruz_cancelled_3y_off(row):
    if row.year + 3 >= row.cancellation_year: return 1
    return 0


financial_ratios['RUZ_cancelled'] = financial_ratios.apply(ruz_cancelled, axis=1)
financial_ratios['RUZ_cancelled_1y_off'] = financial_ratios.apply(ruz_cancelled_1y_off, axis=1)
financial_ratios['RUZ_cancelled_2y_off'] = financial_ratios.apply(ruz_cancelled_2y_off, axis=1)
financial_ratios['RUZ_cancelled_3y_off'] = financial_ratios.apply(ruz_cancelled_3y_off, axis=1)

financial_ratios.shape

for col in ['DPHZ_vat_registration_cancelled',
 'DPHZ_vat_registration_cancelled_1y_off',
 'DPHZ_vat_registration_cancelled_2y_off',
 'DPHZ_vat_registration_cancelled_3y_off',
 'RU_event',
 'RU_event_1y_off',
 'RU_event_2y_off',
 'RU_event_3y_off',
 'RUZ_cancelled',
 'RUZ_cancelled_1y_off',
 'RUZ_cancelled_2y_off',
 'RUZ_cancelled_3y_off']:
    financial_ratios[col] = financial_ratios[col].fillna(0)



financial_ratios.to_csv('../../DATA/MODEL/transformed_set.csv', index = False)

entity_issues_indicators = entity_issues_indicators.query('entity_ico > 0')[['entity_ico', 'year'] + sorted(list(entity_issues_indicators.columns)[2:])].sort_values(['entity_ico', 'year'])

entity_issues_indicators.to_csv('../../DATA/TRANSFORM/issues_indicators.csv', index = False)
